Added leaderboard menu
It can be accessed from the main menu or from the pause menu.
Currently it is only shown on the screen, in the next commit i will make it show the actual leaderboard and let the player save it's data.

+++ Chicken Invaders/Chicken Invaders/Player.h
@@ -79,7 +79,7 @@ private:
 	int m_animRowFrame = 0;
 
 	//Score variables, I may put them the game manager later
-	long long m_score = 0;
+	long m_score = 0;
 	sf::Text m_scoreText;
 
 	sf::Sprite m_ui_rocket;
